+++ CHANGELOG.md
@@ -29,6 +29,12 @@
 
 
 
+* PAPI
+  * Added property name in the error message.
+
+
+
+